Backed out 3 changesets (bug 1459785) for causing https://bugzilla.mozilla.org/show_bug.cgi?id=1464089 a=backout
Backed out changeset 88675b68241a (bug 1459785)
Backed out changeset b7c91a6f1b0a (bug 1459785)
Backed out changeset 21af8dc00aa8 (bug 1459785)

UltraBlame original commit: ff8505d177b9fcba44b040ccd9b6bb709e238e84

+++ gfx/angle/checkout/src/libANGLE/renderer/d3d/BufferD3D.cpp
@@ -160,11 +160,10 @@ void BufferD3D::invalidateStaticData(const gl::Context *context)
 }
 
 
-void BufferD3D::promoteStaticUsage(const gl::Context *context, size_t dataSize)
+void BufferD3D::promoteStaticUsage(const gl::Context *context, int dataSize)
 {
     if (mUsage == D3DBufferUsage::DYNAMIC)
     {
-        
         mUnmodifiedDataUse += dataSize;
 
         if (mUnmodifiedDataUse > 3 * getSize())
